+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Fix: 絵文字管理画面で一部の絵文字が表示されない問題を修正
 
 ### Server
+- Enhance: `/stats` APIの重い処理をスキップできるようにするオプションを追加
+  - 処理がスキップされた部分の値は`0`となります
 - Fix: ユーザーのプロフィール画面をアドレス入力などで直接表示した際に概要タブの描画に失敗する問題の修正( #15032 )
 - Fix: 起動前の疎通チェックが機能しなくなっていた問題を修正  
   (Cherry-picked from https://activitypub.software/TransFem-org/Sharkey/-/merge_requests/737)